api/requirements.txt
appdirs==1.4.3
argon2-cffi==19.1.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.3
Django==2.2
django-filter==2.1.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
Markdown==3.1
MarkupSafe==1.1.1
msgpack-python==0.5.6
openapi-codec==1.3.2
packaging==19.0
Pillow==6.0.0
pip-review==1.0
psycopg2-binary==2.8.2
pycparser==2.19
PyJWT==1.7.1
pyparsing==2.4.0
python-magic==0.4.15
pytz==2019.1
requests==2.21.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.24.2
